Move calculateReadingTime function to blog directory and implement reading time calculation
